requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.6
joblib==1.1.0
python-dotenv==0.19.2
werkzeug==2.0.3
email-validator==1.1.3